requirements.txt
torch==2.1.0
torchvision==0.16.0
streamlit==1.29.0
Pillow==10.0.0
PyYAML==6.0.1
pandas==2.1.0
scikit-learn==1.3.0
tqdm==4.65.0